// SIMS-Front/src/app/models/oswiadczenieDto.ts
export interface oswiadczenieDto {

   opiekunI: string;
   opiekunN: string;
   opiekunMail: string;
   opiekunTel: string;
   studentDuties: string;


  // constructor(id: number = null, opiekunI: string, opiekunN: string, opiekunMail: string, opiekunTel: string) {
  //   this.id = id;
  //   this.opiekunI = opiekunI;
  //   this.opiekunN = opiekunN;
  //   this.opiekunMail = opiekunMail;
  //   this.opiekunTel = opiekunTel;
  // }

}